[google_maps_flutter] Clean Java test, consolidate Marker example. (#4400)
* Do not mock internal class
* Merge "Marker" examples into a single page.
