+++ packages/google_maps_flutter/google_maps_flutter/example/lib/main.dart
@@ -5,7 +5,6 @@
 // ignore_for_file: public_member_api_docs
 
 import 'package:flutter/material.dart';
-import 'package:google_maps_flutter_example/drag_marker.dart';
 import 'package:google_maps_flutter_example/lite_mode.dart';
 import 'animate_camera.dart';
 import 'map_click.dart';
@@ -35,7 +34,6 @@ final List<GoogleMapExampleAppPage> _allPages = <GoogleMapExampleAppPage>[
   PlacePolylinePage(),
   PlacePolygonPage(),
   PlaceCirclePage(),
-  DragMarkerPage(),
   PaddingPage(),
   SnapshotPage(),
   LiteModePage(),
